fixes for checking alerts start and end dates
