+++ src/utils/filterByOpenNow.js
@@ -1,5 +1,4 @@
-const { zonedTimeToUtc, utcToZonedTime, format } = require('date-fns-tz');
-var parseISO = require('date-fns/parseISO');
+import checkAlertsOpenStatus from './checkAlertsOpenStatus';
 
 function filterByOpenNow(locations) {
   const tz = 'America/New_York';
@@ -22,54 +21,7 @@ function filterByOpenNow(locations) {
 
   return locations.reduce((accumlator, location) => {
     // Alerts
-    const alerts = location._embedded.alerts;
-    let alertsOpenStatus;
-    // Check for any alerts.
-    if (alerts === undefined || alerts.length === 0) {
-      // No alerts, so set this status to true.
-      // Other checks below will determine if the location is open.
-      alertsOpenStatus = true;
-    } else {
-      // We have alerts, so map over them.
-      alerts.map(alert => {
-        // Check if closed_for key exists
-        // If so, set the status to false.
-        if ('closed_for' in alert) {
-          // Compare alert.applies.start + alert.applies.end to today.
-          //
-          // Refinery Format:
-          // 2020-07-08T00:00:00-04:00 -- 2030-07-09T00:00:00-04:00
-          //
-          // @SEE https://github.com/NYPL/locations-app/search?q=applies.start&unscoped_q=applies.start
-          //
-          if (alert.applies.start && alert.applies.end) {
-            // Get today date only
-            const utcToday = utcToZonedTime(new Date(), tz);
-            const todayFormatted = format(utcToday, 'yyyy-MM-dd', { timeZone: tz });
-            console.log('todayFormatted: ' + todayFormatted);
-
-            // Get start day
-            // We strip off incorrect offset added.
-            const startDay = format(parseISO(alert.applies.start.replace('-04:00', '')), 'yyyy-MM-dd', { timeZone: tz });
-            console.log('startDay: ' + startDay);
-
-            // Get end day
-            // We strip off incorrect offset added.
-            const endDay = format(parseISO(alert.applies.end.replace('-04:00', '')), 'yyyy-MM-dd', { timeZone: tz });
-            console.log('endDay: ' + endDay);
-
-            // Compare startDay, endDay against today
-            if (todayFormatted != startDay && todayFormatted <= endDay) {
-              alertsOpenStatus = false;
-            } else {
-              alertsOpenStatus = true;
-            }
-          }
-        } else {
-          alertsOpenStatus = true;
-        }
-      });
-    }
+    const alertsOpenStatus = checkAlertsOpenStatus(location);
 
     location.hours.regular.map(hoursItem => {
       // Find today in weekly hours.
